# specs/openapi.yaml
openapi: "3.0.3"
info:
  title: Asistente Ciudadano — API
  description: >
    API del asistente conversacional de trámites públicos. Ver specs/03-api.md
    para reglas transversales.
  version: "1.0.0"
  contact:
    name: Equipo Asistente Ciudadano
servers:
  - url: /api/v1
paths:
  /chat:
    post:
      summary: Enviar un mensaje al asistente
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 1000
                history:
                  type: array
                  maxItems: 20
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      content:
                        type: string
                        maxLength: 2000
      responses:
        "200":
          description: Respuesta del asistente
          content:
            application/json:
              schema:
                type: object
                required: [reply, sources, mode]
                properties:
                  reply:
                    type: string
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        nombre: { type: string }
                  mode:
                    type: string
                    enum: [ai, catalog]
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /tramites:
    get:
      summary: Listar trámites del catálogo
      operationId: listTramites
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
          description: Filtro de texto libre
      responses:
        "200":
          description: Listado resumido
          content:
            application/json:
              schema:
                type: object
                properties:
                  tramites:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        nombre: { type: string }
                        descripcion: { type: string }
                        categoria: { type: string }
                        modalidad: { type: string }
  /tramites/{id}:
    get:
      summary: Ficha completa de un trámite
      operationId: getTramite
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Ficha del trámite
          content:
            application/json:
              schema:
                type: object
        "404": { $ref: "#/components/responses/NotFound" }
  /feedback:
    post:
      summary: Registrar feedback sobre una respuesta
      operationId: feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [helpful]
              properties:
                helpful: { type: boolean }
                comment:
                  type: string
                  maxLength: 500
      responses:
        "204": { description: Registrado }
        "400": { $ref: "#/components/responses/BadRequest" }
  /health:
    get:
      summary: Liveness
      operationId: health
      responses:
        "200":
          description: El proceso está vivo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  version: { type: string }
  /ready:
    get:
      summary: Readiness
      operationId: ready
      responses:
        "200": { description: Listo para recibir tráfico }
        "503": { description: No listo }
components:
  responses:
    BadRequest:
      description: Entrada inválida
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Recurso inexistente
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Límite de solicitudes excedido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }